server/requirements.txt
anyio==4.5.2
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2025.4.26
click==8.1.8
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.95.1
gunicorn==23.0.0
h11==0.16.0
httpcore==0.17.3
httpx==0.24.0
idna==3.10
jinja2==3.1.3
numpy==1.24.4
openai==1.78.0
packaging==25.0
pip==19.2.3
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.0
rich==13.7.0
setuptools==41.2.0
sniffio==1.3.1
starlette==0.26.1
typer==0.9.0
typing-extensions==4.13.2
uvicorn==0.22.0